quick restructure ?
