--- src/component/div/menuView/MenuView.vue
@@ -1,13 +1,0 @@
-<template>
-
-</template>
-
-<script>
-export default {
-}
-
-</script>
-
-<style lang="scss" scoped>
-
-</style>
