+++ src/component/header/HeaderComponent.vue
@@ -2,7 +2,7 @@
   <header>
     <img src="" alt="">
     <div class="message">
-
+        test
     </div>
   </header>
 </template>
