+++ src/main.js
@@ -1,11 +1,11 @@
 import { createApp } from 'vue';
 import App from './App.vue';
 
-import GTIButton from './component/shared/GTIButton.vue';
-import GTITable from './component/shared/GTITable.vue';
+import MyButton from './component/shared/MyButton.vue';
+import MyTable from './component/shared/MyTable.vue';
 
 const app = createApp(App);
 
-app.component("GTIButton", GTIButton);
-app.component("GTITable", GTITable);
+app.component("MyButton", MyButton);
+app.component("MyTable", MyTable);
 app.mount('#app');
