+++ src/views/dataView/StatsView.vue
