+++ src/views/menuView/AddIntPointView.vue
