+++ src/views/menuView/MenuView.vue
@@ -1,5 +1,5 @@
 <template>
-    
+  
 </template>
 
 <script>
@@ -9,5 +9,5 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-    
+
 </style>
